Resolve PR review feedback
- Block empty string values in global property validation
- Reduce max property value length from 500 to 128 chars
- Lift global property keys query to QueryFiltersSelector
- Add console.warn for invalid property value types in analytics script
- Omit global_properties from payload when empty
- Update docs to reflect new value length limit
